backend/requirements/common.txt
asgiref==3.5.2
autopep8==1.6.0
Django==3.2.4
django-cors-headers==3.13.0
django-environ==0.9.0
django_debug_toolbar==3.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
Pillow==9.3.0
pycodestyle==2.8.0
PyJWT==1.7.1
pytz==2022.1
sqlparse==0.4.2
toml==0.10.2
psycopg2==2.9.1